Make it possible to find spicyz from zkg again.
Since about zeek-4.1 the version of zkg distributed with Zeek injects
binary dirs from installed packages into the `PATH` used during e.g.,
`build_command`. We previously removed support for earlier versions by
relying on the fact that `spicyz` was always in `PATH`, but add back
support for finding `spicyz` from installed `zkg` packages as a fallback
if it was not already found in `PATH`.